+++ packages/blade/src/components/Button/BaseButton/__tests__/BaseButton.web.test.tsx
@@ -33,6 +33,7 @@ describe('<BaseButton />', () => {
       }
     }
   });
+
   it('should render xsmall size button', () => {
     const buttonText = 'Pay Now';
     const { container } = renderWithTheme(<BaseButton size="xsmall">{buttonText}</BaseButton>);
@@ -399,11 +400,12 @@ describe('<BaseButton />', () => {
       );
     };
 
-    const { getByText } = renderWithTheme(<ButtonLoading />);
+    const { getByText, container } = renderWithTheme(<ButtonLoading />);
     const loadingButton = getByText(/toggle loading/i);
 
     await user.click(loadingButton);
     expect(getByText(/Started loading/i)).toBeInTheDocument();
+    expect(container).toMatchSnapshot();
     await user.click(loadingButton);
     expect(getByText(/Stopped loading/i)).toBeInTheDocument();
   });