+++ packages/blade/src/components/Input/BaseInput/StyledBaseInput.native.tsx
@@ -137,11 +137,6 @@ export const StyledBaseInput = React.forwardRef<TextInput, StyledBaseInputProps>
         // source: https://reactnative.dev/docs/textinput/#keyboardtype
         keyboardType={KeyboardTypeToNativeValuesMap[keyboardType]}
         returnKeyType={keyboardReturnKeyType}
-        textContentType={
-          autoCompleteSuggestionType
-            ? autoCompleteSuggestionTypeIOS[autoCompleteSuggestionType]
-            : undefined
-        }
         autoCompleteType={
           autoCompleteSuggestionType
             ? (autoCompleteSuggestionTypeAndroid[
@@ -150,6 +145,11 @@ export const StyledBaseInput = React.forwardRef<TextInput, StyledBaseInputProps>
             : undefined
         }
         isTextArea={isTextArea}
+        textContentType={
+          autoCompleteSuggestionType
+            ? autoCompleteSuggestionTypeIOS[autoCompleteSuggestionType]
+            : undefined
+        }
         {...props}
         {...accessibilityProps}
       />