+++ packages/blade/src/components/Input/BaseInput/StyledBaseInput.web.tsx
@@ -6,7 +6,7 @@ import type { StyledBaseInputProps } from './StyledBaseInput.d';
 import getTextStyles from '~components/Typography/Text/getTextStyles';
 
 const StyledBaseNativeInput = styled.input<
-  Omit<StyledBaseInputProps, 'accessibilityProps' | 'setCurrentInteraction'>
+  Omit<StyledBaseInputProps, 'accessibilityProps' | 'setCurrentInteraction' | 'type'>
 >((props) => ({
   ...getBaseInputStyles({
     isDisabled: props.disabled,
@@ -71,6 +71,7 @@ export const StyledBaseInput = React.forwardRef<HTMLInputElement, StyledBaseInpu
       accessibilityProps,
       setCurrentInteraction,
       numberOfLines,
+      type,
       ...props
     },
     ref,
@@ -78,6 +79,7 @@ export const StyledBaseInput = React.forwardRef<HTMLInputElement, StyledBaseInpu
     return (
       <StyledBaseNativeInput
         ref={ref}
+        type={type === 'telephone' ? 'tel' : type}
         disabled={isDisabled}
         required={isRequired}
         maxLength={maxCharacters}